portable shield added

+++ src/main/java/Controller/GameMenuController.java
@@ -46,7 +46,6 @@ public class GameMenuController {
     }
 
     public boolean nextTurn() {
-        //TODO: deal with whose turn is it
         //TODO: check if king is alive or not
         //TODO turns--
         Database.increaseTurnsPassed();
@@ -177,16 +176,21 @@ public class GameMenuController {
             for(int y = startY - range; y < startY + range + 1; y++) {
                 if(!Utils.CheckMapCell.validationOfX(x) || !Utils.CheckMapCell.validationOfY(y)) continue;
 
-                for (Person person : Database.getCurrentMapGame().getMapCellByCoordinates(x, y).getPeople()) {
-                    if(!soldier.getOwner().equals(person.getOwner()) && person.getHp() > 0) {
-                        person.changeHp(-soldier.getAttackRating());
-                        if(person instanceof King && person.getHp() <= 0) {
-                            soldier.getOwner().getEmpire().increaseNumberOfKingsKilled();
-                            destroyEmpire((King)person);
+                AttackToolsAndMethods attackTool = Database.getCurrentMapGame().getMapCellByCoordinates(x, y).getAttackToolsAndMethods();
+                if(attackTool != null && attackTool.getName().equals("portable shield") &&
+                        !attackTool.getOwner().equals(Database.getCurrentUser()) && attackTool.getHp() > 0)
+                    attackTool.changeHp(-soldier.getAttackRating());
+                else
+                    for (Person person : Database.getCurrentMapGame().getMapCellByCoordinates(x, y).getPeople()) {
+                        if(!soldier.getOwner().equals(person.getOwner()) && person.getHp() > 0) {
+                            person.changeHp(-soldier.getAttackRating());
+                            if(person instanceof King && person.getHp() <= 0) {
+                                soldier.getOwner().getEmpire().increaseNumberOfKingsKilled();
+                                destroyEmpire((King)person);
+                            }
+                            break outerLoop;
                         }
-                        break outerLoop;
                     }
-                }
             }
         }
     }
@@ -253,7 +257,8 @@ public class GameMenuController {
                         for (MapCellItems mapCellItem : Database.getCurrentMapGame().getMapCellByCoordinates(x, y).getMapCellItems()) {
                             if(mapCellItem instanceof Wall && !soldier.getOwner().equals(mapCellItem.getOwner()))
                                 ((Wall) mapCellItem).changeHp(-soldier.getAttackRating());
-                        }                    }
+                        }
+                    }
                 }
         }
     }
@@ -291,7 +296,7 @@ public class GameMenuController {
         }
     }
 
-    //TODO: Siege Tower and Portable Shield?
+    //TODO: Siege Tower
 
     public void applyDamageByAttackToolsAndMethods() {
         for (MapCell mapCell : Database.getCurrentMapGame().getMapCells()) {
@@ -316,16 +321,21 @@ public class GameMenuController {
                     break outerLoop;
                 }
 
-                for (Person person : Database.getCurrentMapGame().getMapCellByCoordinates(x, y).getPeople()) {
-                    if(!attackToolsAndMethods.getOwner().equals(person.getOwner()) && person.getHp() > 0) {
-                        person.changeHp(-attackToolsAndMethods.getDamage());
-                        if(person instanceof King && person.getHp() <= 0) {
-                            attackToolsAndMethods.getOwner().getEmpire().increaseNumberOfKingsKilled();
-                            destroyEmpire((King)person);
+                AttackToolsAndMethods attackTool = Database.getCurrentMapGame().getMapCellByCoordinates(x, y).getAttackToolsAndMethods();
+                if(attackTool != null && attackTool.getName().equals("portable shield") &&
+                        !attackTool.getOwner().equals(Database.getCurrentUser()) && attackTool.getHp() > 0)
+                    attackTool.changeHp(-attackToolsAndMethods.getDamage());
+                else
+                    for (Person person : Database.getCurrentMapGame().getMapCellByCoordinates(x, y).getPeople()) {
+                        if(!attackToolsAndMethods.getOwner().equals(person.getOwner()) && person.getHp() > 0) {
+                            person.changeHp(-attackToolsAndMethods.getDamage());
+                            if(person instanceof King && person.getHp() <= 0) {
+                                attackToolsAndMethods.getOwner().getEmpire().increaseNumberOfKingsKilled();
+                                destroyEmpire((King)person);
+                            }
+                            break outerLoop;
                         }
-                        break outerLoop;
                     }
-                }
             }
         }
     }
